pyproject.toml
[tool.poetry]
name = "django-handleref"
version = "0.6.0.1"
description = "django object tracking"
readme = "README.md"
repository = "https://github.com/20c/django-handleref"
authors = ["20C <code@20c.com>"]
license = "Apache-2.0"

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]

packages = [{ include = "django_handleref", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.6.1,<4.0"

[tool.poetry.dev-dependencies]
# test
pytest = ">=6.0.1"
pytest-django = ">=3.8.0"
pytest-cov = "*"
pytest-pythonpath = "*"

# lint
bandit = "^1.6.2"
black = { version = ">=20", allow-prereleases = true }
isort = "^5.7.0"
flake8 = "^3.8.4"
pre-commit = "^2.13.0"
pyupgrade = "^2.19.4"

# docs
markdown = "*"
markdown-include = ">=0.5,<1"
mkdocs = ">=1.0.0,<2.0.0"

# ctl
ctl = "^1.0.0"
jinja2 = "^2.11.2"
tmpl = "^0.3.0"
twine = "^3.3.0"

[tool.poetry.plugins."markdown.extensions"]
pymdgen = "pymdgen.md:Extension"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
